samme koden delt i forskjellige filer for å holde oversikten
